backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
openpyxl==3.1.2
pandas==2.2.2
celery==5.4.0
redis==5.0.4
python-dotenv==1.0.1
pydantic[email]==2.7.1
pydantic-settings==2.2.1
sendgrid==6.11.0
python-docx==1.1.2
reportlab==4.1.0
docx2pdf==0.1.8
Pillow==10.3.0
httpx==0.27.0
tenacity==8.3.0
loguru==0.7.2
pytest==8.2.0
pytest-asyncio==0.23.6
